Use event in webhook api instead of changesets (#50)
Not all webhooks are interested in changesets, but may have the need for changed branches. To support this, we change the extension API so that it no longer provides the changesets, but the event itself.

In addition this has the benefit, that the changesets do not have to be computed for webhooks that are not interested in them.