Enhance API to allow better dto handling (#51)
Adds more interface methods to provide own dtos and preprocess configurations before they are stored. This is necessary for example to support summy passwords in the frontend. To do this, ids are added to each configuration to be able to match changed configurations and with this to compare and handle changes before the new configurations are stored.